Make some refactors and implements and use create-user-policy.
