+++ src/casl/policies/index.ts
@@ -2,16 +2,13 @@ import { Type } from '@nestjs/common';
 
 import { PolicyHandler } from './policy-handler.interface';
 
-import { EditPostHandler } from './edit-post-policy.handler';
-import { CreatePostHandler } from './create-post-policy.handler';
-import { RemovePostHandler } from './remove-post-policy.handler';
+import { postPolicies } from './posts';
+import { userPolicies } from './users';
 
-export * from './edit-post-policy.handler';
-export * from './create-post-policy.handler';
-export * from './remove-post-policy.handler'
+export * from './posts';
+export * from './users';
 
 export const policies: Type<PolicyHandler>[] = [
-  EditPostHandler,
-  CreatePostHandler,
-  RemovePostHandler
+  ...postPolicies,
+  ...userPolicies
 ]
