+++ src/casl/policies/posts/create-post-policy.handler.ts
@@ -4,7 +4,7 @@ import { Post } from '@Posts/entities/post.entity';
 import { AppAbility } from '@Acl/casl-ability.factory';
 import { Action } from '@Acl/enums/action.enum';
 
-import { PolicyHandler } from './policy-handler.interface';
+import { PolicyHandler } from '../policy-handler.interface';
 
 export class CreatePostHandler implements PolicyHandler {
