+++ src/casl/providers/posts/create-post-policy.provider.ts
@@ -1,6 +1,6 @@
 
 import { Provider, Scope } from '@nestjs/common';
 
-import { CreatePostHandler } from '@Acl/policies/create-post-policy.handler';
+import { CreatePostHandler } from '@Acl/policies/posts/create-post-policy.handler';
 
 export const CreatePostPolicyProvider: Provider = CreatePostHandler;
