+++ src/casl/providers/posts/remove-post-policy.provider.ts
@@ -3,7 +3,7 @@ import { Provider } from '@nestjs/common';
 import { REQUEST } from '@nestjs/core';
 import { Request } from 'express';
 
-import { RemovePostHandler } from '@Acl/policies/remove-post-policy.handler';
+import { RemovePostHandler } from '@Acl/policies/posts/remove-post-policy.handler';
 
 export const RemovePostPolicyProvider: Provider = {
   provide: RemovePostHandler,
