+++ src/casl/providers/users/create-user-policy.provider.ts
@@ -1,0 +1,6 @@
+
+import { Provider } from '@nestjs/common';
+
+import { CreateUserHandler } from '@Acl/policies/users/create-user-policy.handler';
+
+export const CreateUserPolicyProvider: Provider = CreateUserHandler;
