+++ src/casl/providers/users/index.ts
@@ -1,0 +1,8 @@
+
+import { Provider } from '@nestjs/common';
+
+import { CreateUserPolicyProvider } from './create-user-policy.provider';
+
+export const userPolicyProviders: Provider[] = [
+  CreateUserPolicyProvider
+]
